# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CoffeeShop API
  description: |
    A RESTful API for managing coffee shop orders.
    
    ## Features
    - Browse available drinks with size restrictions
    - Create and manage customer orders
    - Business rules enforce valid drink/size combinations
    
    ## Authentication
    This demo API does not require authentication. In production, 
    JWT or OAuth2 would be implemented.
  version: 1.0.0
  contact:
    name: API Support
    email: support@coffeeshop.example

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Health
    description: API health check endpoints
  - name: Drinks
    description: Available drinks and their configurations
  - name: Orders
    description: Customer order management

paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns the health status of the API and its dependencies
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: API is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/drinks:
    get:
      tags:
        - Drinks
      summary: List all drinks
      description: Returns all available drinks with their configurations
      operationId: listDrinks
      responses:
        '200':
          description: List of drinks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Drink'
                  count:
                    type: integer
                    example: 4

  /api/v1/drinks/{id}:
    get:
      tags:
        - Drinks
      summary: Get a drink by ID
      description: Returns a single drink by its ID
      operationId: getDrink
      parameters:
        - name: id
          in: path
          required: true
          description: Drink ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Drink details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Drink'
        '404':
          description: Drink not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/orders:
    get:
      tags:
        - Orders
      summary: List all orders
      description: Returns paginated list of orders
      operationId: listOrders
      parameters:
        - name: limit
          in: query
          description: Maximum number of orders to return (1-100)
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of orders to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
    
    post:
      tags:
        - Orders
      summary: Create a new order
      description: |
        Creates a new order with one or more items.
        
        Each item must specify a valid drink_id and size combination.
        The API validates that the requested size is allowed for each drink type.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              simple:
                summary: Simple order
                value:
                  customer_name: "John Doe"
                  items:
                    - drink_id: 1
                      size: "small"
                      quantity: 1
              multiple_items:
                summary: Order with multiple items
                value:
                  customer_name: "Jane Doe"
                  items:
                    - drink_id: 2
                      size: "medium"
                      quantity: 2
                      cup_text: "Jane"
                    - drink_id: 4
                      size: "large"
                      quantity: 1
                  notes: "Extra hot please"
      responses:
        '201':
          description: Order created successfully
          headers:
            Location:
              description: URL of the created order
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/orders/{id}:
    get:
      tags:
        - Orders
      summary: Get an order by ID
      description: Returns a single order with all its items
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Orders
      summary: Update an order
      description: Updates order details (customer name, status, notes)
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Orders
      summary: Delete an order
      description: Permanently deletes an order and all its items
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
      responses:
        '204':
          description: Order deleted successfully
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
        timestamp:
          type: string
          format: date-time
        version:
          type: string
        checks:
          type: object
          properties:
            database:
              type: object
              properties:
                status:
                  type: string
                  enum: [up, down]
                message:
                  type: string

    Drink:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Espresso"
        slug:
          type: string
          example: "espresso"
        type:
          type: string
          enum: [coffee, tea]
          example: "coffee"
        base_price:
          type: number
          format: float
          example: 2.50
        has_milk:
          type: boolean
          example: false
        allowed_sizes:
          type: array
          items:
            type: string
            enum: [small, medium, large]
          example: ["small"]
        components:
          type: array
          items:
            type: string
          example: ["shot of coffee"]

    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        customer_name:
          type: string
          example: "John Doe"
        status:
          type: string
          enum: [pending, preparing, ready, completed, cancelled]
          example: "pending"
        notes:
          type: string
          nullable: true
          example: "Extra hot"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total:
          type: number
          format: float
          example: 5.00
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    OrderItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        drink_id:
          type: integer
          example: 1
        drink_name:
          type: string
          example: "Espresso"
        size:
          type: string
          enum: [small, medium, large]
          example: "small"
        quantity:
          type: integer
          example: 1
        cup_text:
          type: string
          nullable: true
          example: "John"
        price:
          type: number
          format: float
          example: 2.50
        subtotal:
          type: number
          format: float
          example: 2.50

    CreateOrderRequest:
      type: object
      required:
        - customer_name
        - items
      properties:
        customer_name:
          type: string
          minLength: 1
          example: "John Doe"
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - drink_id
              - size
            properties:
              drink_id:
                type: integer
                example: 1
              size:
                type: string
                enum: [small, medium, large]
                example: "small"
              quantity:
                type: integer
                minimum: 1
                maximum: 10
                default: 1
              cup_text:
                type: string
                nullable: true
                example: "John"
        notes:
          type: string
          nullable: true
          example: "No sugar please"

    UpdateOrderRequest:
      type: object
      properties:
        customer_name:
          type: string
          minLength: 1
        status:
          type: string
          enum: [pending, preparing, ready, completed, cancelled]
        notes:
          type: string
          nullable: true

    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0

    Error:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
          example: "Resource not found"
        details:
          type: object
          nullable: true